==> 6 GAPIT-and-FarmCPU.R <==
####GAPIT 
###imported all the data need for GAPIT
myY <- read.csv("data/myYimputedSNP19.csv")
myGD <- read.csv("data/myGDimputedSNP19.csv")
myGM <- read.csv("data/myGMimputedSNP19.csv")
myQ<- read.csv("data/myQimputedSNP19.csv")
###this one for OWA
myY <- read.csv("data/myYimputedSNP19OWA.csv")
str(myY)
myGD <- read.csv("data/myGDimputedSNP19OWA.csv")
str(myGD)
myGM <- read.csv("data/myGMimputedSNP19OWA.csv")
str(myGM)
myQ<- read.csv("data/myQimputedSNP19OWA.csv")
str(myQ)
###install the packaged need for GAPIT
source("http://www.bioconductor.org/biocLite.R")
biocLite("multtest")
install.packages("gplots")
install.packages("LDheatmap")
install.packages("genetics")
install.packages("ape")
install.packages("EMMREML")
install.packages("scatterplot3d") #The downloaded link at: http://cran.r-project.org/package=scatterplot3d
library(multtest)
library(gplots)
library(LDheatmap)
library(genetics)
library(ape)
library(EMMREML)
library(compiler) #this library is already installed in R
library("scatterplot3d")
source("http://zzlab.net/GAPIT/gapit_functions.txt")
source("http://zzlab.net/GAPIT/emma.txt")

####Basic Scenario with Model selection in order to find the optimal number of PCs for different traits. 
setwd("~/Documents/R-corde for miscanthus project/Miscanthus/Result GAPIT1/modelselection")
setwd("C:/Users/Admin/Desktop/Miscanthus/Miscanthus/Result GAPIT1/modelselection")

myGAPIT <- GAPIT(
  Y=myY,
  GD=myGD,
  GM=myGM,
  PCA.total=3, 
  Model.selection = TRUE
)
###through the model selection, the optimal number of PCs is=0, expect the trait of FNMain, PCs=1, 
###for convienence, I set up the PC=0 for all of the traits

####Basic Scenario using PCs=0 without incoporating CV:Population Structure
setwd("~/Documents/R-corde for miscanthus project/Miscanthus/Result GAPIT1/ImputedSNPNOCV")
setwd("C:/Users/Admin/Desktop/Miscanthus/Miscanthus/Result GAPIT1/ImputedSNPNOCV")
myGAPIT <- GAPIT(
  Y=myY,
  GD=myGD,
  GM=myGM
)

####Basic Scenario using PCs=0 with incoporate CV:Population Structure
setwd("~/Documents/R-corde for miscanthus project/Miscanthus/Result GAPIT1/ImputedSNPCV")
setwd("C:/Users/Admin/Desktop/Miscanthus/Miscanthus/Result GAPIT1/ImputedSNPCV")
myGAPIT <- GAPIT(
  Y=myY,
  GD=myGD,
  GM=myGM,
  CV=myQ
)

####Enhanced Compression without CV
setwd("~/Documents/R-corde for miscanthus project/Miscanthus/Result GAPIT1/ECMMLNOCV")
setwd("C:/Users/Admin/Desktop/Miscanthus/Miscanthus/Result GAPIT1/ECMMLNOCV")
myGAPIT <- GAPIT(
  Y=myY,
  GD=myGD,
  GM=myGM,
  kinship.cluster=c("average", "complete", "ward"), kinship.group=c("Mean", "Max"), 
  group.from=100,
  group.to=32338, 
  group.by=10
)

####Enhanced Compression with CV
  setwd("~/Documents/R-corde for miscanthus project/Miscanthus/Result GAPIT1/ECMMLCV")
  setwd("C:/Users/Admin/Desktop/Miscanthus/Miscanthus/Result GAPIT1/ECMMLCV")
  myGAPIT <- GAPIT(
    Y=myY,
    GD=myGD,
    GM=myGM,
    CV=myQ,
    kinship.cluster=c("average", "complete", "ward"), kinship.group=c("Mean", "Max"), 
    group.from=100,
    group.to=32338, 
    group.by=10
  )
  
###########"GLM","MLM","MLMM","FarmCPU" without CV
  setwd("~/Documents/R-corde for miscanthus project/Miscanthus/Result GAPIT1/GLMMLMMLMMFarmCPU")
  setwd("C:/Users/Admin/Desktop/Miscanthus/Miscanthus/Result GAPIT1/GLMMLMMLMMFarmCPU")
  myGAPIT <- GAPIT(
    Y=myY,
    GD=myGD,
    GM=myGM,
    model=c("GLM","MLM","MLMM","FarmCPU")
  )
  
###########"GLM","MLM","MLMM","FarmCPU" with CV
  setwd("~/Documents/R-corde for miscanthus project/Miscanthus/Result GAPIT1/GLMMLMMLMMFarmCPUCV")
  setwd("C:/Users/Admin/Desktop/Miscanthus/Miscanthus/Result GAPIT1/GLMMLMMLMMFarmCPUCV")
  myGAPIT <- GAPIT(
    Y=myY,
    GD=myGD,
    GM=myGM,
    CV=myQ, 
    model=c("GLM","MLM","MLMM","FarmCPU")
  )
  
####ALL of R codes below using Bottom as Supper and changed different top. but without incorporating CV, 
####becasue i saw the example does not use the PCA, but CV in the manual 
##########Using MLM SUPPER with CV
  setwd("~/Documents/R-corde for miscanthus project/Miscanthus/Result GAPIT1/MLMSUPPER")
  setwd("C:/Users/Admin/Desktop/Miscanthus/Miscanthus/Result GAPIT1/MLMSUPPER")
  myGAPIT <- GAPIT(
    Y=myY,
    GD=myGD,
    GM=myGM,
    CV=myQ,
    sangwich.top="MLM", #options are GLM,MLM,CMLM, FaST and SUPER 
    sangwich.bottom="SUPER", #options are GLM,MLM,CMLM, FaST and SUPER 
    LD=0.1
  )
  
###########Using GLM SUPPER with CV
  setwd("~/Documents/R-corde for miscanthus project/Miscanthus/Result GAPIT1/GLMSUPPER")
  setwd("C:/Users/Admin/Desktop/Miscanthus/Miscanthus/Result GAPIT1/GLMSUPPER")
  myGAPIT <- GAPIT(
    Y=myY,
    GD=myGD,
    GM=myGM,
    CV=myQ,
    sangwich.top="GLM", #options are GLM,MLM,CMLM, FaST and SUPER 
    sangwich.bottom="SUPER", #options are GLM,MLM,CMLM, FaST and SUPER 
    LD=0.1
  )
  
###########Using CMLM SUPPER with CV
  setwd("~/Documents/R-corde for miscanthus project/Miscanthus/Result GAPIT1/CMLMSUPPER")
  setwd("C:/Users/Admin/Desktop/Miscanthus/Miscanthus/Result GAPIT1/CMLMSUPPER")
  myGAPIT <- GAPIT(
    Y=myY,
    GD=myGD,
    GM=myGM,
    CV=myQ,
    sangwich.top="CMLM", #options are GLM,MLM,CMLM, FaST and SUPER 
    sangwich.bottom="SUPER", #options are GLM,MLM,CMLM, FaST and SUPER 
    LD=0.1
  )
  
###########Using SUPPER SUPPER with CV
  setwd("~/Documents/R-corde for miscanthus project/Miscanthus/Result GAPIT1/SUPPERSUPPER")
  setwd("C:/Users/Admin/Desktop/Miscanthus/Miscanthus/Result GAPIT1/SUPPERSUPPER")
  myGAPIT <- GAPIT(
    Y=myY,
    GD=myGD,
    GM=myGM,
    CV=myQ,
    sangwich.top="SUPER", #options are GLM,MLM,CMLM, FaST and SUPER 
    sangwich.bottom="SUPER", #options are GLM,MLM,CMLM, FaST and SUPER 
    LD=0.1
  )
